requirements.txt
flask==2.3.3
twilio==8.5.0
pandas==2.0.3
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
supabase==2.0.3
openpyxl==3.1.2
python-dateutil==2.8.2
pydrive==1.3.1
oauth2client==4.1.3 